profile added

+++ routes/web.php
@@ -30,6 +30,9 @@
         Route::get('/profile/{id}', [UserProfileController::class, 'show'])->name('admin.user-profile');
         Route::get('/profile/{id}/profile-edit', [UserProfileController::class, 'edit'])->name('admin.profile-edit');
         Route::put('/profile/{id}/profile-edit', [UserProfileController::class, 'edit'])->name('admin.profile-edit');
+        Route::put('/profile/{id}/profile-update', [UserProfileController::class, 'update'])->name('admin.profile-update');
+        Route::put('/profile/{id}/profile-delete', [UserProfileController::class, 'delete'])->name('admin.profile-delete');
+        Route::delete('/issues/{id}/remove-profile-pic', [UserProfileController::class, 'removeProfilePic'])->name('remove-profile-pic');
 
         Route::get('/project-list', [ProjectController::class, 'index'])->name('admin.project-list');
         Route::get('/project-create', [ProjectController::class, 'create'])->name('admin.project-create');
@@ -48,7 +51,7 @@
         Route::put('/issues/{id}/issue-edit', [IssueController::class, 'edit'])->name('admin.issue-edit');
         Route::put('/issues/{id}/issue-update', [IssueController::class, 'update'])->name('admin.issue-update');
         Route::delete('/issues/{id}/issue-delete', [IssueController::class, 'delete'])->name('admin.issue-delete');
-        Route::delete('/issues/{id}/remove-attachment', [IssueController::class, 'removeAttachment'])->name('remove-attachment');
+        Route::delete('/issues/{id}/remove-attachment', [IssueController::class, 'removeAttachment'])->name('admin.remove-attachment');
 
         Route::get('/user-list', [UserController::class, 'user_list'])->name('admin.user-list');
         Route::get('/user-register', [UserController::class, 'create'])->name('admin.user-register');
@@ -66,6 +69,9 @@
         Route::get('/profile/{id}', [UserProfileController::class, 'show'])->name('user.user-profile');
         Route::get('/profile/{id}/profile-edit', [UserProfileController::class, 'edit'])->name('user.profile-edit');
         Route::put('/profile/{id}/profile-edit', [UserProfileController::class, 'edit'])->name('user.profile-edit');
+        Route::put('/profile/{id}/profile-update', [UserProfileController::class, 'update'])->name('user.profile-update');
+        Route::put('/profile/{id}/profile-delete', [UserProfileController::class, 'delete'])->name('user.profile-delete');
+        Route::delete('/issues/{id}/remove-profile-pic', [UserProfileController::class, 'removeProfilePic'])->name('remove-profile-pic');
 
         Route::get('/project-list', [ProjectController::class, 'index'])->name('user.project-list');
         Route::get('/project-create', [ProjectController::class, 'create'])->name('user.project-create');
@@ -83,7 +89,7 @@
         Route::put('/issues/{id}/issue-edit', [IssueController::class, 'edit'])->name('user.issue-edit');
         Route::put('/issues/{id}/issue-update', [IssueController::class, 'update'])->name('user.issue-update');
         Route::delete('/issues/{id}/issue-delete', [IssueController::class, 'delete'])->name('user.issue-delete');
-        Route::delete('/issues/{id}/remove-attachment', [IssueController::class, 'removeAttachment'])->name('remove-attachment');
+        Route::delete('/issues/{id}/remove-attachment', [IssueController::class, 'removeAttachment'])->name('user.remove-attachment');
 
     });
